requirements.txt
fastapi
uvicorn
python-multipart
httpx
pillow
dotenv
google-genai